+++ CHANGELOG.md
@@ -23,6 +23,10 @@ Some examples, more below in the actual changelog (newer entries are more likely
 
 -->
 
+### Changed
+* lbaas/v1: changed state retriever interface `StateSuccess`, `StateProgressing` & `StateFailure` to `StateOK`, `StatePending` & `StateError` (#185, @marioreggiori)
+* (internal) generic client: common Generic Service interfaces and utils have been extracted into a new shared package (#185, @marioreggiori)
+
 ## [0.4.6] - 2022-10-04
 
 ### Added